data_engineering/pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "data_engineering"
version = "0.0.1"
description = "analysis of dataset"
authors = [
  { name = "Anna Vasilevskaia" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "notebook",
    "numpy",
    "pandas",
    "pandera",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "scikit-learn",
    "tqdm",
    "typer",
    "wget (>=3.2,<4.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "black (>=25.9.0,<26.0.0)",
]
requires-python = "==3.13.5"


[tool.ruff]
line-length = 99
src = ["data_engineering"]
include = ["pyproject.toml", "data_engineering/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["data_engineering"]
force-sort-within-sections = true

